Avoid using (u)intmax_t in methods
It is not guaranteed that a type encoding for it exists.

FossilOrigin-Name: d74f244ab7a1ca9855c2415281bea54a923ce23dadac30af7bc48a6f75a7c9fd

+++ src/OFFileManager.h
@@ -578,7 +578,7 @@ OF_SUBCLASSING_RESTRICTED
  *
  * Raises an @ref OFUndefinedKeyException if the key is missing.
  */
-@property (readonly, nonatomic) uintmax_t fileSize;
+@property (readonly, nonatomic) unsigned long long fileSize;
 
 /*!
  * @brief The @ref of_file_attribute_key_type key from the dictionary.
